requirements.txt
bokeh >= 3.5.2, < 3.6

confusable_homoglyphs >= 3.3.1, < 3.4.0

coverage >= 5.5, < 6.0

Django >= 4.2.22, < 5.0

djangorestframework >= 3.15.0, < 3.16.0

djangorestframework-guardian2

django-admin-interface >= 0.28.3

django-cors-headers>=4.4.0,<4.5.0

django-filter >= 23.5, < 24.0

django-health-check >= 3.18.0, < 3.19.0

django-model-utils >= 4.3.1, < 4.4.0

django-notifications-hq >= 1.8.3, < 2.0

django-rest-knox >= 5.0.2, < 5.1.0

docutils >= 0.20.1

drf-spectacular == 0.25.1

flake8==7.1.1

gunicorn >= 23.0.0, < 23.1.0

isort >= 5.11.0, < 5.12.0

PyJWT>=2.0.0

psycopg2-binary >= 2.9, < 3.0

requests >= 2.32.4, < 2.33.0

setuptools == 68.0.0
